requirements.txt
PySide6==6.9.1
PyInstaller==6.14.2
Pillow==11.3.0
twitchio==3.0.1
requests==2.32.4
aiohttp==3.12.14
python-socketio==5.13.0
